Refactor/frontend modify

+++ backend/src/repositories/facultyRepository.ts
@@ -1,10 +1,10 @@
+import Student from "@models/student";
 import Faculty, { IFacultyDocument } from "../models/faculty";
 
 export const createFaculty = async (
-  name: string,
-  code: string
+  name: string
 ): Promise<IFacultyDocument> => {
-  return await Faculty.create({ name, code });
+  return await Faculty.create({ facultyName: name });
 };
 
 export const getAllFaculties = async (): Promise<IFacultyDocument[]> => {
@@ -13,12 +13,11 @@ export const getAllFaculties = async (): Promise<IFacultyDocument[]> => {
 
 export const updateFaculty = async (
   id: string,
-  name: string,
-  code: string
+  newName: string
 ): Promise<IFacultyDocument | null> => {
   return await Faculty.findByIdAndUpdate(
     id,
-    { name, code },
+    { facultyName: newName },
     { new: true }
   ).exec();
 };
@@ -26,9 +25,13 @@ export const updateFaculty = async (
 export const deleteFaculty = async (
   id: string
 ): Promise<IFacultyDocument | null> => {
-  return await Faculty.findByIdAndDelete(id).exec();
-};
+  // Check if there are any students associated with this faculty
+  const studentCount = await Student.countDocuments({ faculty: id }).exec();
+  if (studentCount > 0) {
+    throw new Error(
+      `Cannot delete faculty with id "${id}" because it is referenced by ${studentCount} student(s).`
+    );
+  }
 
-export const getFacultyByCode = async (code: string) => {
-  return await Faculty.findOne({ code }).exec(); // Find faculty by code
+  return await Faculty.findByIdAndDelete(id).exec();
 };
